move convert_type_maybe to operator/convop.py
