+++ spy/vm/typechecker.py
@@ -4,13 +4,13 @@ from spy import ast
 from spy.irgen.symtable import Symbol, Color
 from spy.errors import (SPyTypeError, SPyNameError, maybe_plural)
 from spy.location import Loc
+from spy.vm.modules.operator.convop import convert_type_maybe
 from spy.vm.object import W_Object, W_Type
 from spy.vm.opimpl import W_OpImpl, W_OpArg
 from spy.vm.list import W_List, make_oparg_list
-from spy.vm.function import W_FuncType, W_ASTFunc, W_Func
+from spy.vm.function import W_ASTFunc, W_Func
 from spy.vm.b import B
 from spy.vm.modules.operator import OP, OP_from_token
-from spy.vm.modules.jsffi import JSFFI
 from spy.vm.modules.types import W_TypeDef
 from spy.util import magic_dispatch
 if TYPE_CHECKING:
@@ -556,63 +556,3 @@ def _call_error_wrong_argcount(
         err.add('note', 'function defined here', def_loc)
     raise err
 
-def convert_type_maybe(
-        vm: 'SPyVM',
-        wop_x: W_OpArg,
-        w_exp: W_Type
-) -> Optional[W_Func]:
-    """
-    Check whether the given W_OpArg is compatible with the expected type:
-
-      - if it's the same type (no conversion needed), return None
-
-      - if it can be converted, return a w_func which does the conversion
-
-      - if the types are not compatible, raise SPyTypeError. In this case,
-        the caller can catch the error, add extra info and re-raise.
-    """
-    w_got = wop_x.w_static_type
-    if vm.issubclass(w_got, w_exp):
-        # nothing to do
-        return None
-
-    # try to see whether we can apply a type conversion
-    if vm.issubclass(w_exp, w_got):
-        # this handles two separate cases:
-        #   - upcasts, e.g. object->i32: in this case we just do a typecheck
-        #   - dynamic->*: in this case we SHOULD do actual conversions, but at
-        #                 the moment we don't so we conflate the two cases
-        #                 into one
-        w_from_dynamic_T = vm.call(OP.w_from_dynamic, [w_exp])
-        return w_from_dynamic_T
-
-    # XXX move this dictionary somewhere else
-    converters_w = {
-        (B.w_i32, B.w_f64): OP.w_i32_to_f64,
-        (B.w_i32, B.w_bool): OP.w_i32_to_bool
-    }
-    key = (w_got, w_exp)
-    w_conv = converters_w.get(key)
-    if w_conv is not None:
-        return w_conv
-
-    if w_exp is JSFFI.w_JsRef:
-        if w_conv := convert_JsRef_maybe(w_got, w_exp):
-            return w_conv
-
-    # mismatched types
-    err = SPyTypeError('mismatched types')
-    got = w_got.fqn.human_name
-    exp = w_exp.fqn.human_name
-    err.add('error', f'expected `{exp}`, got `{got}`', loc=wop_x.loc)
-    raise err
-
-
-def convert_JsRef_maybe(w_got: W_Type, w_exp: W_Type) -> Optional[W_Func]:
-    if w_got is B.w_str:
-        return JSFFI.w_js_string
-    elif w_got is B.w_i32:
-        return JSFFI.w_js_i32
-    elif isinstance(w_got, W_FuncType):
-        assert w_got == W_FuncType.parse('def() -> void')
-        return JSFFI.w_js_wrap_func
